Remove the Dataset heading in the debug output
The ehrQL debug command no longer displays the dataset output by
default.

Typically, the debug command will not write anything to stdout
anymore, but I've left it in the output because it might at some
point (i.e. we intend to support print() statements in the dataset
definition in some form, and those will write to stdout by default.)