Add the ehrQL debug command to the Run button menu
We also enable the command and the menu item only when the
editor is active and the language is python. (Although we still have to
do the checks that the editor exists in the ts file, otherwise the
linter complains).
The "navigation" group is a special group that ensures the command will
be at the top of the options (and therefore the default).
Also removes the status bar button.