Only update the isEhrql config when the active text editor changes
Avoids re-running the isEhrql checking on every keystroke. It does mean
that a new dataset definition won't get picked up as soon as the word
ehrql is typed, but this is probably relatively rare - most users will
be starting from the research template, which has a stub dataset
definition
file in it already.