Switch to using `spawn` rather than `exec`.
Because:

a) exec uses shell, which is less secure
b) it simplifies the quoting of arguments
c) we can control the stdout/stderr

We use this to merge the stdout/stderr into a single in-order stream, as
a array of strings.

This will allow better formatting of plain text output, whether it came
from stderr or stdout.