Build output as a string instead of a list
The output from child_process.spawn has a finite size; if we're trying
to show a table with several hundred rows, it is batched, which means
that we end up with the html table string split. When we then join it
again for the html panel display, we get line breaks in the middle of
the html, which breaks the formatting.

We now build the output as a single string, instead of a list of
strings. I have verified that non-table output print() statements do
get correctly displayed on new lines.