Dump the ehrql example-data as dummy tables
If the dummy tables directory doesn't already exists, dump the ehrQL
example-data into the expected location. This means that a new user
won't encounter just an unfathomable vscode error saying that dummy
tables don't exist.  However, the example data is limited and created
with the ehrQL tutorial in mind; it won't contain all the tables or
fields that all users might want to use. So we print a warning to the
output panel to that effect.